requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
Pillow>=9.5.0
datasets>=2.12.0
scikit-learn>=1.2.0
bitsandbytes>=0.40.0
accelerate>=0.20.0
openai-clip @ git+https://github.com/openai/CLIP.git
sentencepiece>=0.1.99
protobuf>=3.20.0
safetensors>=0.3.1 